FastApi/server/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
psycopg2-binary
email-validator
bcrypt
python-jose[cryptography]
python-multipart
slowapi
